Extracted tableViewDataSource and delegate from Countries table view into a helper class
